Move CommandBatch[Form] into DevTools.RuntimeDeploy.Infrastructure
Relocate the deploy-batch console window pair from the shared DevTools.Libs
into the only project that uses it. This is the first concrete step toward
dropping the DevTools.Libs project reference from DevTools.RuntimeDeploy.

Moves:
- CommandBatch.cs
- CommandBatchForm.cs / .Designer.cs / .resx

All four files use git-mv so commit history is preserved as renames.
Namespace changes from DevTools.Libs to DevTools.RuntimeDeploy.Infrastructure.

Folded into this commit (step 2 of the plan, "replace-filecopy"):
- Replace the FileInfoExtension.CopyToAsync calls in
  CommandBatch.CopyFileCommandAsync with synchronous
  File.Copy(src, dst, overwrite: true). The async stream-copy helper was
  swallowing exceptions into a returned status string; switching to
  File.Copy lets failures throw naturally. The "OK." log line moves out
  of the conditional since the synchronous Copy either succeeds or
  throws.
- Delete src/WinFormsToolLib/FileInfoExtension.cs (no other consumers).

The two steps had to land together because moving CommandBatch broke
its access to the formerly-internal FileInfoExtension.CopyToAsync.

DevTools.Libs.csproj cleanup:
- Drop WarpToolkit.WinForms and WarpToolkit.WinForms.Tooling package
  references (no longer used; only DebugListener's
  WarpToolkit.Desktop.Async.Collections remains).
- Drop the obsolete <Compile Update="CommandBatchForm.cs" /> entry.

DevTools.RuntimeDeploy.csproj:
- Add <Compile Update="Infrastructure\CommandBatchForm.cs" /> so the
  designer recognises the partial-class pair (no SubType needed since
  the form is purely code-generated; Designer support is preserved by
  the .Designer.cs neighbour).

DeployRuntimeView.cs picks up the new namespace via an added
'using DevTools.RuntimeDeploy.Infrastructure;'.
