+++ src/DevTools.RuntimeDeploy/Infrastructure/CommandBatchForm.resx
